Added simple command line client
Added a simple command line client that uses SyncBot.

Argument processing is a little brutal but works, would like to add the
ability to read P4 environment variables in the future.

Also changed the interface to SyncBot. Now instead of explicitly passing
a path and a changelist, you only pass path. This allows a user to use
all of the P4 path syntax.